requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.19
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2 